requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Pillow == 7.0.0

PyMuPDF == 1.18.4

nltk == 3.5

pdf2image == 1.14.0

pytesseract == 0.3.6
